rename file and use local watch vs passed state prop
